refactor(tests): split fake repository search
